# apps/web/src/lib/contracts/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hubris-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": ["crate::api::settings"],
        "summary": "GET /api/settings",
        "operationId": "get_settings",
        "responses": {
          "200": {
            "description": "Current settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsState"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": ["crate::api::settings"],
        "summary": "PUT /api/settings — full replace",
        "operationId": "put_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsState"
                }
              }
            }
          },
          "409": {
            "description": "Settings file is invalid on disk"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": ["crate::api::settings"],
        "summary": "PATCH /api/settings — partial update",
        "operationId": "patch_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings patched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsState"
                }
              }
            }
          },
          "409": {
            "description": "Settings file is invalid on disk"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppearanceSettings": {
        "type": "object",
        "properties": {
          "colorScheme": {
            "$ref": "#/components/schemas/ColorScheme"
          },
          "darkTheme": {
            "type": "string"
          },
          "lightTheme": {
            "type": "string"
          }
        }
      },
      "AppearanceSettingsPatch": {
        "type": "object",
        "properties": {
          "colorScheme": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ColorScheme"
              }
            ]
          },
          "darkTheme": {
            "type": ["string", "null"]
          },
          "lightTheme": {
            "type": ["string", "null"]
          }
        }
      },
...
      "ChatSettings": {
        "type": "object",
        "description": "Chat settings owned by the backend.",
        "required": ["idleTimeoutMinutes"],
        "properties": {
          "idleTimeoutMinutes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ChatSettingsPatch": {
        "type": "object",
        "properties": {
          "idleTimeoutMinutes": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ColorScheme": {
        "type": "string",
        "enum": ["auto", "light", "dark"]
      },
...
      "EditorSettings": {
        "type": "object",
        "properties": {
          "darkEditorTheme": {
            "type": "string"
          },
          "lightEditorTheme": {
            "type": "string"
          }
        }
      },
      "EditorSettingsPatch": {
        "type": "object",
        "properties": {
          "darkEditorTheme": {
            "type": ["string", "null"]
          },
          "lightEditorTheme": {
            "type": ["string", "null"]
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "appearance": {
            "$ref": "#/components/schemas/AppearanceSettings"
          },
          "chat": {
            "$ref": "#/components/schemas/ChatSettings"
          },
          "editor": {
            "$ref": "#/components/schemas/EditorSettings"
          },
          "terminal": {
            "$ref": "#/components/schemas/TerminalSettings"
          },
          "vscode": {
            "$ref": "#/components/schemas/VscodeSettings"
          },
          "worktree": {
            "$ref": "#/components/schemas/WorktreeSettings"
          }
        }
      },
      "SettingsPatch": {
        "type": "object",
        "properties": {
          "appearance": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AppearanceSettingsPatch"
              }
            ]
          },
          "chat": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChatSettingsPatch"
              }
            ]
          },
          "editor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EditorSettingsPatch"
              }
            ]
          },
          "terminal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TerminalSettingsPatch"
              }
            ]
          },
          "vscode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VscodeSettingsPatch"
              }
            ]
          },
          "worktree": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorktreeSettingsPatch"
              }
            ]
          }
        }
      },
      "SettingsState": {
        "type": "object",
        "required": ["settings", "generation", "status"],
        "properties": {
          "generation": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/Settings"
          },
          "status": {
            "$ref": "#/components/schemas/SettingsStatus"
          }
        }
      },
      "SettingsStatus": {
        "type": "object",
        "required": ["kind", "writesBlocked"],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SettingsStatusKind"
          },
          "message": {
            "type": ["string", "null"]
          },
          "writesBlocked": {
            "type": "boolean"
          }
        }
      },
      "SettingsStatusKind": {
        "type": "string",
        "enum": ["ok", "invalidFile"]
      },
...
      "TerminalFontSource": {
        "type": "string",
        "enum": ["default", "system", "bundled"]
      },
      "TerminalSettings": {
        "type": "object",
        "required": [
          "fontSource",
          "systemFontFamily",
          "bundledFont",
          "fontSize",
          "smartTabNaming",
          "escapeSequenceTitles",
          "sendKeybindingsToShell",
          "clientScrollbackRows",
          "serverScrollbackBytes"
        ],
        "properties": {
          "bundledFont": {
            "type": "string"
          },
          "clientScrollbackRows": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "escapeSequenceTitles": {
            "type": "boolean"
          },
          "fontSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "fontSource": {
            "$ref": "#/components/schemas/TerminalFontSource"
          },
          "sendKeybindingsToShell": {
            "type": "boolean"
          },
          "serverScrollbackBytes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "smartTabNaming": {
            "type": "boolean"
          },
          "systemFontFamily": {
            "type": "string"
          }
        }
      },
      "TerminalSettingsPatch": {
        "type": "object",
        "properties": {
          "bundledFont": {
            "type": ["string", "null"]
          },
          "clientScrollbackRows": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "escapeSequenceTitles": {
            "type": ["boolean", "null"]
          },
          "fontSize": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "fontSource": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TerminalFontSource"
              }
            ]
          },
          "sendKeybindingsToShell": {
            "type": ["boolean", "null"]
          },
          "serverScrollbackBytes": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "smartTabNaming": {
            "type": ["boolean", "null"]
          },
          "systemFontFamily": {
            "type": ["string", "null"]
          }
        }
      },
...
      "VscodeRuntimeKind": {
        "type": "string",
        "enum": ["vscodeCli", "codeServer"]
      },
...
      "VscodeSettings": {
        "type": "object",
        "properties": {
          "runtime": {
            "$ref": "#/components/schemas/VscodeRuntimeKind"
          }
        }
      },
      "VscodeSettingsPatch": {
        "type": "object",
        "properties": {
          "runtime": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VscodeRuntimeKind"
              }
            ]
          }
        }
      },
...
      "WorktreeLocationMode": {
        "type": "string",
        "enum": ["dataDir", "repoLocalDotHubris"]
      },
...
      "WorktreeSettings": {
        "type": "object",
        "properties": {
          "locationMode": {
            "$ref": "#/components/schemas/WorktreeLocationMode"
          }
        }
      },
      "WorktreeSettingsPatch": {
        "type": "object",
        "properties": {
          "locationMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorktreeLocationMode"
              }
            ]
          }
        }
      },
...
    }
  },
...
}